+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New version of event manager interface (`IEventManager`).
 - New instantiatable event manager implementing `IEventManager`.
+- Unit tests for new instantiatable event manager.
 
 ### Changed